+++ react/components/PONumber.tsx
@@ -3,6 +3,7 @@ import { useIntl } from 'react-intl'
 import { Input } from 'vtex.styleguide'
 
 import { useOrderFormCustom } from '../hooks'
+import { CompleteOrderForm } from '../typings'
 import {
   B2B_CHECKOUT_CUSTOM_APP_ID,
   B2B_CHECKOUT_CUSTOM_APP_MAJOR,
@@ -32,7 +33,7 @@ export function PONumber() {
               },
             ],
           },
-    })
+    } as CompleteOrderForm)
   }
 
   return (