+++ react/hooks/useClearCart.ts
@@ -2,21 +2,18 @@ import { useMutation } from '@tanstack/react-query'
 
 import { useCheckoutB2BContext } from '../CheckoutB2BContext'
 import { apiRequest } from '../services'
-import type { ApiResponse, WithToast } from '../typings'
-import { useOrderFormCustom, UseOrderFormReturn } from './useOrderFormCustom'
+import type { ApiResponse, CompleteOrderForm, WithToast } from '../typings'
+import { useOrderFormCustom } from './useOrderFormCustom'
 
 export function useClearCart(showToast: WithToast['showToast']) {
   const { orderForm, setOrderForm } = useOrderFormCustom()
   const { setPending } = useCheckoutB2BContext()
 
-  const { mutate, isLoading } = useMutation<
-    UseOrderFormReturn['orderForm'],
-    Error
-  >({
+  const { mutate, isLoading } = useMutation<CompleteOrderForm, Error>({
     mutationFn: async () => {
       setPending(true)
 
-      return apiRequest<UseOrderFormReturn['orderForm'] & ApiResponse>(
+      return apiRequest<CompleteOrderForm & ApiResponse>(
         `/api/checkout/pub/orderForm/${orderForm.id}/items/removeAll`,
         'POST',
         {}